requirements/base.txt
Django>=3.0,<4.0
djangorestframework>=3.10,<4.0
psycopg2-binary>=2.8,<3.0
django-filter>=2.0,<3.0
drf-yasg>=1.20,<2.0
python-dotenv>=0.19,<1.0